Reorganize/rename EE MatView modules.
